+++ projects/ixocreate/ngx-admin/src/lib/pipes/pipes.module.ts
@@ -12,7 +12,7 @@ const COMPONENTS = [
   IxoDateTimePipe,
   IxoFileSizePipe,
   IxoNumberPipe,
-  IxoTimePipe,
+  IxoTimePipe
 ];
 
 @NgModule({