+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/admin-frontend/compare/0.1.11...develop)
+## [Unreleased](https://github.com/ixocreate/admin-frontend/compare/0.1.12...develop)
+
+## [v0.1.12 (2019-08-22)](https://github.com/ixocreate/admin-frontend/compare/0.1.11...0.1.12)
+### Changed
+- Added a preview button and a live updating preview screen
+- You can now resize the preview panel
+- Added buttons at the bottom of the preview panel that sets screen sizes automatically 
 
 ## [v0.1.11 (2019-08-22)](https://github.com/ixocreate/admin-frontend/compare/0.1.10...0.1.11)
 ### Changed